PC port: Add PC_BUILD ifdefs for decomp integration
Adds platform abstraction throughout the decomp codebase:
- GBI: _PC_GFX_STASH for 64-bit pointer preservation in display lists
- Boot: boot_main returns instead of looping on PC
- Main loop: Tracing and PC-safe control flow
- State machines: Skip ROM DMA, load assets from PNG files
- Sprites/models: Guard N64-specific initialization
- background_gfx: Runtime DL init for 64-bit (gs* -> g* macros)

+++ src/43F0.c
@@ -64,7 +64,7 @@ HeapNode* _heap_create(HeapNode* addr, u32 size) {
     if (size < 32) {
         return (HeapNode*)-1;
     } else {
-        HeapNode* heapNode = (HeapNode*)ALIGN16((u32)addr);
+        HeapNode* heapNode = (HeapNode*)ALIGN16((uintptr_t)addr);
 
         size -= ((u8*)heapNode - (u8*)addr);
         heapNode->next = nullptr;
